continuous read to the max

+++ include/i2c_fpga/i2c.hpp
@@ -37,7 +37,7 @@ public:
 
 	void write(uint8_t i2cAddr, uint32_t data, uint8_t number_of_bytes);
 	uint32_t read(uint8_t i2cAddr, uint8_t reg, uint8_t number_of_bytes);
-    void read_continuous(uint8_t i2cAddr, uint8_t number_of_bytes, vector<uint32_t> &data);
+    void read_continuous(uint8_t i2cAddr, uint8_t number_of_bytes, vector<uint8_t> &data);
 };
 
 #define IORD(base,reg) (*(((volatile int32_t*)base)+reg))
